Relocate CloudFoundry support
