+++ azure-spring-boot/src/main/java/com/microsoft/azure/spring/cloundfoundry/environment/VcapServiceType.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See LICENSE in the project root for
  * license information.
  */
-package com.microsoft.azure.cloudfoundry.service;
+package com.microsoft.azure.spring.cloundfoundry.environment;
 
 public enum VcapServiceType {
     AZURE_REDISCACHE("azure-rediscache"), AZURE_DOCUMENTDB("azure-documentdb"), AZURE_STORAGE(
